+++ libastfri-uml/astfri-uml/impl/PlantUMLOutputter.cpp
@@ -8,19 +8,19 @@ std::string PlantUMLOutputter::assemble_param(VarStruct p) {
     std::string result;
 
     if (p.isIndirect_)
-        p.type_ += this->config_->indirectIndicator_;
+        p.type_ += this->config_->indirectIndicator;
     result += TypeConvention::get_string(
         p.type_,
         p.name_,
-        this->config_->separator_,
-        this->config_->typeConvention_
+        this->config_->separator,
+        this->config_->typeConvention
     );
 
     return result;
 }
 
 void PlantUMLOutputter::open(const ClassStruct& cs) {
-    if (this->config_->handleNamespaces_ && ! cs.namespace_.empty()) {
+    if (this->config_->handleNamespaces && ! cs.namespace_.empty()) {
         this->outputString_ += cs.namespace_;
     }
     else {
@@ -43,16 +43,16 @@ void PlantUMLOutputter::open(const ClassStruct& cs) {
 
 void PlantUMLOutputter::apply_style_from_config() {
     std::string style = "<style>\n";
-    style += "classDiagram {\nBackGroundColor " + this->config_->diagramBG_ + "\n}\n";
-    style += "class {\nBackGroundColor " + this->config_->elementBG_ + "\n";
-    style += "LineColor " + this->config_->elementBorder_ + "\n";
-    style += "FontColor " + this->config_->fontColor_ + "\n" + "}\n";
-    style += "arrow {\nLineColor " + this->config_->arrowColor_ + "\n}\n";
+    style += "classDiagram {\nBackGroundColor " + this->config_->bgDiagram + "\n}\n";
+    style += "class {\nBackGroundColor " + this->config_->bgElement + "\n";
+    style += "LineColor " + this->config_->elementBorder + "\n";
+    style += "FontColor " + this->config_->fontColor + "\n" + "}\n";
+    style += "arrow {\nLineColor " + this->config_->arrowColor + "\n}\n";
     style += "</style>\n";
-    if (this->config_->handleNamespaces_) {
-        style += "set separator " + this->config_->namespaceSeparator_ + "\n";
+    if (this->config_->handleNamespaces) {
+        style += "set separator " + this->config_->namespaceSeparator + "\n";
     }
-    if (! this->config_->drawAccessModIcons_)
+    if (! this->config_->drawIcons)
         style += "skinparam classAttributeIconSize 0\n";
     this->outputString_ = style + this->outputString_;
 }
@@ -103,20 +103,20 @@ void PlantUMLOutputter::close_user_type() {
 
 void PlantUMLOutputter::add_data_member(VarStruct v) {
     if (v.isIndirect_)
-        v.type_ += this->config_->indirectIndicator_;
-    this->outputString_ += this->config_->accessPrefix_[(int)v.accessMod_]
+        v.type_ += this->config_->indirectIndicator;
+    this->outputString_ += this->access_mod_to_char(v.accessMod_)
                          + TypeConvention::get_string(
                                v.type_,
                                v.name_,
-                               this->config_->separator_,
-                               this->config_->typeConvention_
+                               this->config_->separator,
+                               this->config_->typeConvention
                          )
                          + "\n";
 }
 
 void PlantUMLOutputter::add_function_member(MethodStruct m) {
     if (m.returnIsIndirect_)
-        m.retType_ += this->config_->indirectIndicator_;
+        m.retType_ += this->config_->indirectIndicator;
     std::string header = m.name_ + "(";
     size_t index       = 0;
     for (VarStruct p : m.params_) {
@@ -128,19 +128,19 @@ void PlantUMLOutputter::add_function_member(MethodStruct m) {
     }
     header += ")";
 
-    this->outputString_ += this->config_->accessPrefix_[(int)m.accessMod_]
+    this->outputString_ += this->access_mod_to_char(m.accessMod_)
                          + TypeConvention::get_string(
                                m.retType_,
                                header,
-                               this->config_->separator_,
-                               this->config_->typeConvention_
+                               this->config_->separator,
+                               this->config_->typeConvention
                          )
                          + "\n";
 }
 
 void PlantUMLOutputter::add_constructor(ConstructorStruct c) {
     std::string header;
-    if (config_->innerView_) {
+    if (config_->innerView) {
         header += c.class_ + "(";
     }
     else {
@@ -157,20 +157,20 @@ void PlantUMLOutputter::add_constructor(ConstructorStruct c) {
     }
     header += ")";
 
-    this->outputString_ += this->config_->accessPrefix_[(int)c.accessMod_]
+    this->outputString_ += this->access_mod_to_char(c.accessMod_)
                          + TypeConvention::get_string(
                                c.class_,
                                header,
-                               this->config_->separator_,
-                               this->config_->typeConvention_
+                               this->config_->separator,
+                               this->config_->typeConvention
                          )
                          + "\n";
 }
 
 void PlantUMLOutputter::add_destructor(DestructorStruct d) {
     std::string header = "";
-    header += this->config_->accessPrefix_[(int)AccessModifier::Public];
-    header += this->config_->destructorIndicator_;
+    header += this->access_mod_to_char(AccessModifier::Public);
+    header += this->config_->destructorIndicator;
     header += d.class_;
     header += "()\n";
     this->outputString_ += header;
@@ -178,6 +178,38 @@ void PlantUMLOutputter::add_destructor(DestructorStruct d) {
 
 void PlantUMLOutputter::add_relation(RelationStruct r) {
     this->outputString_
-        += r.to_ + " " + this->config_->relationArrows_[(int)r.type_] + " " + r.from_ + "\n";
+        += r.to_ + " " + this->relation_to_arrow(r.type_) + " " + r.from_ + "\n";
 }
+
+char PlantUMLOutputter::access_mod_to_char(AccessModifier am) const {
+    switch (am) {
+        case AccessModifier::Public:
+            return config_->publicPrefix;
+        case AccessModifier::Private:
+            return config_->privatePrefix;
+        case AccessModifier::Protected:
+            return config_->protectedPrefix;
+        default:
+            throw std::runtime_error(
+                "Unhandled enum value in `PlantUMLOutputter::access_mod_to_char`.");
+    }
+}
+
+std::string PlantUMLOutputter::relation_to_arrow(RelationType rt) const {
+    switch (rt) {
+        case RelationType::ASSOCIATION:
+            return config_->association;
+        case RelationType::COMPOSITION:
+            return config_->composition;
+        case RelationType::EXTENSION:
+            return config_->extension;
+        case RelationType::IMPLEMENTATION:
+            return config_->implementation;
+        default:
+            throw std::runtime_error(
+                "Unhandled enum value in `PlantUMLOutputter::relation_to_arrow`.");
+    }
+}
+
+
 } // namespace astfri::uml